requirements.txt
pandas
scikit-learn
flask
sqlalchemy
pymysql
joblib
